# [Enhancement] Modify the matching rules of generic functions