--- a/fe/fe-core/src/test/java/com/starrocks/catalog/FunctionSetTest.java
+++ b/fe/fe-core/src/test/java/com/starrocks/catalog/FunctionSetTest.java
@@ -261,6 +261,12 @@ public class FunctionSetTest {
         Assert.assertEquals(Type.ARRAY_BIGINT, fn.getReturnType());
         Assert.assertEquals(Type.BIGINT, fn.getArgs()[0]);
 
+        // arrays_overlap
+        argTypes = new Type[] {Type.ARRAY_BIGINT, Type.ARRAY_TINYINT };
+        desc = new Function(new FunctionName("arrays_overlap"), argTypes, Type.BOOLEAN, false);
+        fn = functionSet.getFunction(desc, Function.CompareMode.IS_SUPERTYPE_OF);
+        Assert.assertNotNull(fn);;
+        Assert.assertEquals(fn.functionId, 150216L);
     }
 
     @Test