## Why I'm doing:

## What I'm doing:

After specialization of the prototype function, if it can be strictly matched, replace it with a strictly matched function.

Performance test: 2.7s->1.3s

```
CREATE TABLE `t1` (                                                                                                                                                                                                                                                     
  `c1` bigint(20) NOT NULL COMMENT "",                                                                                                                                                                                                                                            
  `c2` array<bigint(20)> NOT NULL COMMENT "",                                                                                                                                                                                                                                     
  `c3` array<bigint(20)> NOT NULL COMMENT ""                                                                                                                                                                                                                                      
) ENGINE=OLAP                                                                                                                                                                                                                                                                     
DUPLICATE KEY(`c1`)                                                                                                                                                                                                                                                               
DISTRIBUTED BY HASH(`c1`) BUCKETS 1                                                                                                                                                                                                                                               
PROPERTIES (                                                                                                                                                                                                                                                                      
"compression" = "LZ4",                                                                                                                                                                                                                                                            
"fast_schema_evolution" = "true",                                                                                                                                                                                                                                                 
"replicated_storage" = "true",                                                                                                                                                                                                                                                    
"replication_num" = "1"                                                                                                                                                                                                                                                           
);

mysql> select count(*) from t1;                                                                                                                                                                                                                                                   
+----------+                                                                                                                                                                                                                                                                      
| count(*) |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
| 30000000 |                                                                                                                                                                                                                                                                      
+----------+

Before optimization: 2.68s

mysql> select count(*) from t1 where arrays_overlap(c2, [1, 11, 111, 1111, 11111, 111111]);                                                                                                                                                                                       
+----------+                                                                                                                                                                                                                                                                      
| count(*) |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
|     1204 |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
1 row in set (2.68 sec) 

After optimization: 1.31s

mysql> select count(*) from t1 where arrays_overlap(c2, [1, 11, 111, 1111, 11111, 111111]);
+----------+
| count(*) |
+----------+
|     1204 |
+----------+
1 row in set (1.31 sec)
```

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5